feat(cli): improve optional import error (#35239)
# Why

I tried using `react-native-vision-camera` which has a number of
optional features and setup instructions that are a bit hard to follow,
this led to the missing optional import error in Metro:

```
Error: Requiring unknown module "undefined". If you are sure the module exists, try restarting Metro. You may also want to run `yarn` or `npm install`.
```

Obviously `Requiring unknown module "undefined".` is not helpful. This
happens when you have some code like this:

```js
function loadFunky() {
  try {
    return require("@shopify/react-native-skia");
  } catch (e) {
    throw new Error("never thrown for some reason");
  }
}
```

For whatever reason the `never thrown for some reason` is never thrown
for some reason, perhaps it's the error guard that wraps the import
(confirmed it is absolutely because of the error guard) but it's besides
the point. The error message should never say "undefined", it should
have the name of the imported module.

We can partially fix this in development bundles when using our version
of Metro runtime `EXPO_USE_METRO_REQUIRE=1`. This will use the require
hint in the bundle to populate the error message correctly:

```
Error: Requiring unknown module "@shopify/react-native-skia". If you are sure the module exists, try restarting Metro. You may also want to run `yarn` or `npm install`.
```

This also modifies the collect-dependencies code to ensure the import
hint is always included for optional imports so production errors also
have a useful message.

# Test Plan

In a project with some code like this:


```js
function loadFunky() {
  try {
    return require("@shopify/react-native-skia");
  } catch (e) {
    throw new Error("never thrown for some reason");
  }
}
```

Run the project `EXPO_USE_METRO_REQUIRE=1 npx expo start` -> the error
will be helpful now. Not sure if an upstream version of this would be
accepted so merging here first.

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>