feat(metro): add web worker support on web (#34938)
# Why

- @Kudo requested this feature as a slightly better version of just
adding a file to the public folder or creating an inline blob.


<!--
Please describe the motivation for this PR, and link to relevant GitHub
issues, forums posts, or feature requests.
-->

# How

- First I looked at adding module qualifiers like `./my-worker?worker`
but this doesn't work in Metro because Metro only supports transforming
a single copy of any given file based on the file path. I added the
extraction to collect dependencies but wasn't able to get the qualifier
to the transformer or add the qualifiers to the transform cache key.
- Next I added `require.resolve` to add support for getting a processed
URL component for an async dependency but this doesn't work for web
workers because web workers are essentially standalone JS bundles and
need their own copy of the metro runtime in the split bundle.
- Then I added `require.resolveWorker` which would transform into a
string but mark the dependency as being a special type of async chunk
that acts as a standalone entry. In the split chunks algo I made each
worker chunk also get access to its own copy of the premodules and I
removed the worker chunks from deduplication as they wouldn't be sharing
code with the main chunk.
- I also added support for marking `new Worker(new URL('./foo',
window.location.href))` to collect dependencies, but I don't love how
fragile this syntax is for marking a dependency. Would prefer something
that explicitly throws an error when it cannot be marked, whereas this
matcher would just create a different string that doesn't form a full
bundle.
- I ran into issues with HMR because the HMR serializer is different to
the serializer used for everything else, meaning the worker URL would
change when saving a file and use a corrupt URL. Forked the HMR
serializer to get around this.
- Considering using an alternative approach here that sits outside of
Metro, similar to DOM components. Problem with this would be the need to
maintain another forked bundling process.
- Workers don't have HMR built-in and don't work on native. The bundling
is technically supported on native but the native module for Worker is
not. Unclear what treatment we should add to support code sharing here.
- The workers approach I'm currently on requires lazy bundling to be
enabled otherwise it'll throw cryptic errors, wonder if it makes sense
to just forcing it to be enabled.
- Because we're bundling for a different env, we probably also want to
set the `customTransformOptions.environment` to `worker` or something
related.
- Unclear if we can add HMR here.

# Test Plan

- Added web e2e tests in development and production which also test HMR
updates to ensure the HMR serializer does not corrupt the worker import.
- Added mini-metro tests to ensure the splitting and tree shaking
features work as expected.
- Added AST parsing and dependency collection tests.

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>
Co-authored-by: Aman Mittal <amandeepmittal@live.com>