refactor(cli): remove usage of `@expo/rudder-sdk-node` in favor of `FetchClient` (#35299)
# Why

This removes the usage of `@expo/rudder-sdk-node` in favor of our
existing `FetchClient` and `FetchDetachedClient`. Unfortunately, the
Rudder SDK package is outdated and was built for servers, not clients
like the CLI. We've partially moved over from this package to the
`FetchClient` for the `expo start`, `expo run*`, and `expo export`
commands in SDK 52 - and we are happy with the data it provides.

For SDK 53, we want to fully remove the package because it's outdated
and larger than it has to be for the tasks its supposed to do.

# How

- PR #35271 removed the usage of the Rudder package to the `FetchClient`
and is backported to SDK 52
- This PR fully removes the package for SDK 53

# Test Plan

- Run any of the commands
- Telemetry should still be sent if not disabled with
`EXPO_NO_TELEMETRY`

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [ ] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)