[fix][android][tests][updates] fixed complicated solution (#35293)
# Why

In #34363 a bit of an over-engineered solution was added to fix a
failing ErrorRecovery test after we added RN 0.77
(ReactNativeFeatureFlags was implemented a bit differently).

# How

The fix removes the complex solution and adds a much simpler one as
suggested by @Kudo.

# Test Plan

Run Android tests in BareExpo

# Checklist

- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).