chore(expo-asset): Sync changes from `@expo/metro-config` and drop unused dependencies (#35328)
# Why

In #34208, @EvanBacon removed the need for `hashAssetFiles` and the
logic is being moved to `@expo/metro-config`. While we're hoping to
remove this file for SDK 53 entirely, in case we don't and in the
meantime, it'd be great to drop this dependency to make the dependency
graph of canary releases representative of this future removal.

Specifically, the `md5-file` dependency isn't being used in the new
implementation (since it was replaced with a few lines of custom code)
and can be sync'ed back to this file. A comment was left to clarify
this.

It also seems like `invariant` isn't actually used but being specified
as a dependency.

# How

- Sync md5 utility back from `@expo/metro-config` based on #34208 
- Remove `md5-file` dependency
- Remove unused `invariant` dependency

# Test Plan

- Manually tested against an SDK 52 project

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [x] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)