[ios][background-task] Added showing an error when bg tasks are not available
Closes #ENG-15092

If background tasks aren't available we need to show a message to the user trying to explain whats going on.

Usually it will be because the user is trying to run on an iOS simulator or web.

This commit fixes this by both displaying a message over console.error to ensure that it is visible even though we're in an async method and then also throwing an error to stop further execution.