fix(metro-config): Fix tree shaking multiple exports in the same variable declaration. (#35330)
# Why

- As reported by RBI, the library `@amplitude/analytics-browser` has
some syntax that we weren't accounting for in tree-shaking which caused
the output to be corrupted.

```js
export var foo = '...', bar = '...';
```

If `bar` wasn't used, then foo would be removed and take the entire
declaration with it. We now filter on the identifiers and only remove
the export if it's empty.

Also adds fix for renamed imports:

```ts
import foo from './unused';
import { foo as bar } from './bar';
console.log(bar);
```

- `foo` should be removed.
- `bar` should remain.

# Test Plan

- Added a new mini-metro test.
- E2E tests should keep working.

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>