fix(cli): Silence missing favicon file error. (#35357)
# Why

This pull request includes a bug fix to handle missing favicon files
gracefully and updates the changelog to reflect this fix. The changes
are primarily focused on error handling in the `FaviconMiddleware`.

Bug Fixes:

*
[`packages/@expo/cli/src/start/server/middleware/FaviconMiddleware.ts`](diffhunk://#diff-5eaf98dff66c77281db0569bbd3afeacf926fd2ff733f2cefcf18582293d0606R37-R41):
Added a condition to bypass errors when the favicon file is not found
(`ENOENT` error).

Documentation:

*
[`packages/@expo/cli/CHANGELOG.md`](diffhunk://#diff-dced9f7f9eaad1257a4d1786ee6facb5010790cec922eb8662e33a83923c7f7dR18):
Updated the changelog to include the fix for silencing missing favicon
file errors.

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>