[video][ios] Fix the Now Playing notification disappearing after the video is paused (#35273)
# Why

Part of a minor audit of Now Playing for iOS (1/3)
Fixes https://github.com/expo/expo/issues/33785
The Now Playing notification disappears on iOS after the video is paused

# How

The `if` statement in `VideoManager` is too restrictive. We also have to
check if the playback category and mode are correct.
Calling the `NowPlayingManager` from the function also turned out to be
unnecessary after fixes from the next PR are applied.
Moreover the `findAudioMixingMode` would incorrectly return the audio
mixing to be set as `mixWithOthers` while the audio mixing mode was
unset. Those three things caused the notification to disappear after the
video is paused


# Test Plan

Tested in BareExpo on an iPhone 13