[video][ios] Fix a race condition in setting the targets for the Now Playing controls (#35274)
# Why

Part of a minor audit of Now Playing for iOS (2/3)

When the Now Playing notification is disabled and enabled again the
controls are sometimes inactive or only some of them work.

# How

When we disable and re-enable the controls we call
`removeNowPlayingTargets` and then `addNowPlayingTargets` right after
that. I believe that underneath the calls to the command center are
re-dispatched and somehow the individual line call order is mismatched,
which causes the control targets to be added and removed at the same
time. This means that some of the time all of them work correctly, and
some of the time only some or none of them work.

Dispatching the removal and addition of the targets from the main queue
seems to fix the issue.

# Test Plan

Tested in BareExpo on iPhone 13.