[video][ios] Fix a race condition when unregistering multiple players from Now Playing at the same time (#35275)
# Why

Part of a minor audit of Now Playing for iOS (3/3)

When we were having multiple players registered for displaying of the
notification and disabled all of them at the same time the notification
would not disappear.

Currently here is what was happening:
1. Unregister Player1
2. Unregister Player1 has elected Player2 as the new source for the
notification - start the async task that loads the data for the
notification
3. Unregister Player2 - No more players are available - disable the
notification (task is still running)
4. The notification is disabled (this operation is synchronous and fast)
5. Task has finished - Updates the now playing notification with stale
data, also making it impossible to hide the notification
(`NowPlayingManager` "thinks" that the notification is disabled).

This is an edge case that shouldn't really happen in real world too
often, but still worth fixing.

# How

Assigned the task to a variable. The task is cancelled every time a new
player is elected as the notification source (nil included). THanks to
that we will never reach the step 5 in above, which means that the
notification is disabled correctly.

# Test Plan

Tested in Bare Expo on iPhone 13