[ios][location] Added scope to back-/foreground permissions
In the docs we refer to a field called scope on iOS in the LocationPermissionResponse which is not set by the foreground/background permission requestors.

This PR fixes this by setting it. It might not seem super important, but it is documented and we should "do as we say".

Added some tests to verify that this field is set as well.