[ios][location] Clean up NSNull values from NSDict/NSArray
When passing task options to set up the CLLocationManager the options struct may contain keys with undefined (from JS) values - which will end up as NSNull values and crash the app. This can be seen in the API/BackgroundLocation example in BareExpo where some of the option fields are undefined.

This commit fixes this by reading these values in a safer way.