[cli] export:embed should create tmp assets directory if needed  (#35387)
If an app has only the JS bundle, and no other assets, the iOS "Bundle React Native code and images" task will fail for release builds.

Repro:

- Create a new app with the `blank-typescript` template
- Remove the status bar import and markup from `App.tsx`
- Do a clean prebuild
- `npx expo run:ios --configuration Release`

In export:embed, after creating the JS bundle, add a check to see if the assets directory exists, and create it if needed.

- The above repro should succeed in building the release build for iOS
- CI should pass