[ios][background-task] Added showing an error when bg tasks are not available (#35350)
If background tasks aren't available we need to show a message to the
user trying to explain whats going on.

Closes #ENG-15092

Usually it will be because the user is trying to run on an iOS simulator
or web.

This commit fixes this by throwing an error if you try to call
`registerTask` on an iOS device or an unsupported device.

Run in BareExpo on iOS Simulator and add a call to `registerTask` in
`BackgroundTaskScreen.tsx`
Observe that an exception is showed.

- [X] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [X] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>
Co-authored-by: Brent Vatne <brentvatne@gmail.com>