[ios][location] don't restart all services on error (#35478)
When CLLocationManager reports an error in our TaskConsumer we shouldn't
need to restart location updates and monitoring - just handle the error
through the task system. This caused the location manager to spit out
errors over and over again since each error would try to restart -
creating a new error (if the error is persistent, like if location
services are unavailable for a while).

This commit fixes this by removing the restart logic. This is now
similar to how the TaskConsumer for monitoring regions are doing it.

- Run in BareExpo on Simulator and start tracking background locations
(remember to remove `<MapView` from the BackgroundLocationsScreen.tsx to
be able to test)
- Set navigation simulation to "None" (Features -> Location -> "None")
to provoke an error
- Observe that we're not seeing a lot of restart tries in the log
window.

- [x] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [x] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>