[ios][calendar] fix: allow more characters in url encoding (#35034)
# Why

When creating an event with a full URL, the URL in the event is broken.
Example :
- Create an event with the url 'https://google.com'
- The event will have the url 'https%3A%2F%2Fgoogle.com'

# How

I added new allowed characters in the URL encoding (: and /)

# Test Plan

To test this change, create a new event with an URL. Then check for the
created event in the calendar.

Or simply create an Xcode Playground only with the maybeSetUrl function

---------

Co-authored-by: Łukasz Kosmaty <kosmatylukasz@gmail.com>