[store-review][iOS] fix `activationState` must be used from main thread only (#35403)
# Why
Xcode warning in code:
<img width="459" alt="Scherm­afbeelding 2025-03-10 om 15 58 19"
src="https://github.com/user-attachments/assets/eac9796d-919d-4fd2-af22-dfc672004d53"
/>

Also in the logs it shows a warning when executing the requestReview
method.

# How

I made sure that `getForegroundActiveScene()` also get's called on the
main thread.

# Test Plan

Without this change Xcode warns that the code is executed in background
while it should be executed on the main thread.