[android][audio] Fix muting (#35631)
# Why
Correctly handles the volume on android. Determining if the audio is muted should not be done through `isDeviceMuted` on the `ExoPlayer` instance. This refers to the devices global sound setting. It should instead be determined on the the volume of the player.

# How
Update `muted` to use `volume == 0f`. Because users can write to the `muted` property this leads to a complications with the `volume` property. Both are now controlling the same underlying property so we need to store the muted state on the player instance as well as the last non 0 volume to determine what we should be setting the volume to at any given time. 

# Test Plan
Bare-expo. Muting and unmuting works as expected and unmuting restores the volume to the previous level it was at.