[video] Add `onFirstFrameRender` event to the `VideoView` (#35346)
# Why

We don't have an event that should be used when the first frame of the
video has been successfully displayed in the `VideoView`. `onLoad` and
the `readyToPlay` status are called before the first frame is rendered,
therefore they cannot be used to a cover image.

# How

Added`onFirstFrameRender` event to the `VideoView`. This is not a player
event, because on iOS a single player can be attached to multiple views.

# Test Plan

Tested in BareExpo on Android, iOS and Web

Example of cover image usage:


https://github.com/user-attachments/assets/2cfb9d0d-2635-4613-b959-d137742609a1