[video][android] Add `useExoShutter` property to the `VideoView` (#35366)
# Why

The videoView is all black until the first frame is loaded. On iOS the
view is transparent before the first frame is rendered. This is the
default ExoPlayer behavior so we probably want to keep it, but we can
add a prop that allows users to align the behavior with iOS.

Fixes https://github.com/expo/expo/issues/31561

This also allows people to display stuff underneath the controls before
the first frame is rendered


# How

When it comes to `exo_shutter` we can just set it to be transparent when
the prop changes. The `SurfaceView` is a bit more complex. We keep track
of when we want to hide the `SurfaceView` and apply the visibility when
needed. This is necessary, because the `SurfaceView` will be black when
empty instead of being transparent.

On the native side we use our own `onFirstFrameRendered` event to avoid
showing `SurfaceView` layout changes.

# Test Plan

Tested in BareExpo on Android

The first video shows the behaviour with `useExoShutter={true}`
(default), the second one with `useExoShutter={false}`.


[default_shutter.webm](https://github.com/user-attachments/assets/6bc908f3-4978-4177-a6d5-bd6a4225e4d3)


[no_shutter.webm](https://github.com/user-attachments/assets/5d3e2ed6-0e90-40a0-b6aa-88c0d326b0d4)