[router]: Use UNSTABLE_router & add 'dangerouslySingular' prop (#35595)
# Why

Replacement PR to https://github.com/expo/expo/pull/35544

Uses the new `UNSTABLE_router` feature to override the internal router.
This PR minimised the `react-native-screens` stack rearrangement issue
and introduces a new `<Stack.Screen unique />` prop called `unique`,
which is an easier to use `getId` replacement.

# How

## withLayoutContext

When you declare a navigator in a layout, if you provide a `getId`
function that will be passed unaltered to the `<Screen />`. Expo Router
no longer adds a `getId` to every screen. I have marked `getId` as
deprecated in the PR, as it will be replaced with the easier to use
`unique` prop.

## `unique` 

`unique` is the replacement to `getId`. It accepts these values 

```
unique?: 
  | (options: { params?: Record<string, any>, name: string }) => string | undefined
  | string // Alias for () => <value>
  | boolean // Sets the getId as the previous Expo Router getId function
```

> [!NOTE]
> Using `unique` re-introduces the issue that causes
`react-native-screens` to break. This issue has not been resolved! This
PR just changes the default behaviour so people will not be affected. If
people opt into using `getId`/`unique` then they need to manage this bug
themselves.


So you can have a `_layout`

```
<Stack>
  <Stack.Screen name="feed" />
  <Stack.Screen name="[profile]" unique />
</Stack>
```

If the history is

- `/my-profile`

When you `router.push('/feed'`); router.push('/feed')` the the state
will be

- `/my-profile`
- `/feed`
- `/feed`

If you then push `router.push('/my-profile)` the history will be

- `/feed`
- `/feed`
- `/my-profile`

If you then push `router.push('/other-profile')` the history will be

- `/feed`
- `/feed`
- `/my-profile`
- `/other-profile`

This is because we take both the `name` and the route params in account
when calculating uniqueness (same as the old `getId` function). Search
params are not considered.

`unique` can also be used dynamically and does not need to be set in the
`_layout`

For example, passing `unique` to link `<Link href="/feed" unique />`
will cause that link to act as if `/feed` had `unique`. This also works
with the imperative API `router.push('/feed', { unique: true })`.
`unique` here accepts the same options as `<Screen />`

You can also pass a custom `unique` function to `<Link />` that will
check for uniqueness only for **that** `<Link />`

When you use `unique` in a dynamic manner (e.g on a `<Link />`) it will
remove any duplicate "unique" routes. For example, you may have a
history of

- `/feed`
- `/feed`
- `/feed`
- `/my-profile`

If you press `<Link href="/feed" unique />`, your history will now be

- `/my-profile`
- `/feed`

If you `push`, it will always perform this collapsing logic. While
`navigate` will only collapse if it navigates

# Test Plan

> [!IMPORTANT]
> This PR should not modify existing tests! We want to retain the same
behaviour we have now.

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [ ] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)