[image-picker][android] fix `getPendingResultAsync` returning `null`
The `getPendingResultAsync` function's return type is an array of either
`ImagePickerResult | ImagePickerErrorResult`, but its native implementation
sometimes returns `null`:

```kotlin
    AsyncFunction("getPendingResultAsync") Coroutine { ->
      val (bareResult, options) = pendingMediaPickingResult ?: return@Coroutine null
      // ... etc.
```

This can cause JavaScript runtime errors, if code is written to expect an array.

The proposed fix fixes this on the TypeScript side, returning an empty array when
the native function returns `null`, but a fix could also be made on the Kotlin side
(I avoided doing so, mostly, because contributing to the native code looks more
complicated; please alter my change, if this is the preferred approach).