[create-expo] Support aliasing and deprecating examples (#35717)
# Why

We have a bunch of examples in https://github.com/expo/examples that
we'd like to remove, rename, or merge. Adding this functionality allows
us to keep the repo more organized and easy to browse, without breaking
any existing documentation (or at least, not without a useful error
message).

# How

- I added
[**meta.json**](https://github.com/expo/examples/blob/master/meta.json)
to https://github.com/expo/examples
- We fetch that from `create-expo` and use it to determine whether to
remap to another example (alias) or throw a meaningful error
(deprecation)

## Alias

```
create-expod -e with-tab-navigation
The with-tab-navigation example has been renamed to with-react-navigation.
Creating an Expo project using the with-react-navigation example.

? What is your app named? › my-app
```

You can just alias it and the message above will be displayed, or you
can add a `message` too in order to provide more context:

```
The with-tab-navigation example has been renamed to with-react-navigation.
This example has been replaced with a more complete navigation example, which also includes tabs.
Creating an Expo project using the with-react-navigation example.

? What is your app named? › my-app
```

## Deprecation

```
create-expod -e with-formik
Error: with-formik is no longer available. Formik is a plain-JavaScript library that does not require any specific configuration related to React Native or Expo. Follow the Formik documentation: https://formik.org/

You can also refer to the outdated example code in examples git repository history, if it is useful: https://github.com/expo/examples/tree/7a5145833837b9e3c72ebe381c06925b7fc5daf6/with-formik
```

Note: in this case, we can optionally provide `message` (the "Formik is
a plain-JavaScript library" part), and an `outdatedExampleHref` (the
https://github.com/expo/examples/tree/7a5145833837b9e3c72ebe381c06925b7fc5daf6/with-formik
link), so we can guide people in the right direction.

# Test Plan

Ran it locally, verified it all looks good. 

# Checklist

- [x] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>