Update @shopify/react-native-skia to version v2.0.0-next.2 (#35698)
# Why

This commit updates @shopify/react-native-skia from 1.5.0 ->
v2.0.0-next.2

Closes ENG-15319

# How

- Updated package.json in expo-go + ncl
- Updated bundledNativeModules.json
- Updated yarn.lock
- Updated podfile.lock for expo-go

(BareExpo isn't built with RNSkia)

# Test Plan

Run Expo Go locally:
- [X] Tested on Expo Go iOS
- [X] Tested on Expo Go Android

# Notes:

~I found an
[issue](https://github.com/Shopify/react-native-skia/issues/2974) in
their repo about not being compatible with React 19 and RN 0.78 yet
(which causes it to crash on Expo Go on main even before updating). This
is fixed in a next
[release](https://github.com/Shopify/react-native-skia/releases/tag/v2.0.0-next.1)
in RNSkia - but it requires Android SDK 26 and might not be backwards
compatible with RN < 0.78 when released. Seems to be OK for ExpoGo
(except for the min SDK restriction?) - is this ok for us to upgrade
to?~

New RNSKIA version 2.0.0-next.2 reverted the minSdk change.