[ios][xcode] fix compile errors with xcode 16.3
https://developer.apple.com/documentation/xcode-release-notes/xcode-16_3-release-notes

Some stuff that makes BareExpo/Expo-go fail to build:

- libc++ no longer supports std::allocator<const T>
- The base template for std::char_traits has been removed.

In addition it seems like TARGET_OS_SIMULATOR is not available in Swift anymore (change to use targetEnvironment(simulator))

This commit fixes the issues related to the above by changing the use of `TARGET_OS_SIMULATOR` in swift code to `#if targetEnvironment(simulator)` instead.

In addition the error with `std::allocator<const T>` is fixed in EXGLNativeContext.h where we `used a std::set<const std::string>`.

Tested by building both BareExpo and ExpoGo on iOS with new XCode 16.3