[media-library][1/2] Make moving assets unnecessary -  Add `album` parameter to `createAssetAsync` (#35686)
# Why

Currently to create an asset inside of an album the users have to 
1. Create the asset
2. Move the asset to an existing album

On android this poses a problem, because moving the asset is equivalent
to copying it to new location and deleting it from the old one. The
deletion operation requires a manual permission confirmation from the
user every time we do it.
Alternatively the users can copy the asset instead of moving, but in
that case the default photo library gets polluted with our assets.

# How

Add `album` property to `createAssetAsync` - The asset will be created
in the target album instead of being created in the default album.
This change is fully backward compatible. If no `album` was provided we
keep the old behaviour.

# Test Plan

Tested in BareExpo on Android and iOS.