[background-task] handling TaskManager errors (#35734)
# Why

When trying to register an unregistered task, or unregsiter a task that
isn't registered, the app will crash with a native failure that causes
hit to hard crash.

# How

This commit fixes this by adding a test to see if it is registered or
unregistered before calling the native code.

# Test Plan

Test in BareExpo - try to call register on a task that is already
registered.

# Checklist

- [x] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [x] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>