fix(cli): force react navigation to use import instead of require. (#35904)
# Why

The new metro package exports syntax chooses which condition to use
based on if the module is imported using `import` or `require` which
means you can import multiple copies of the same package. This can lead
to issues where a node module has a React context that isn't shared.
Many libraries will potentially have this issue it's not entirely clear
what the best move is here, potentially locking an import condition per
bundle based on the first instance (which would require tracking across
async bundles in dev), warning, or forcing everyone not to use require
anymore (which is seemingly what the plan was).

People still use `require` syntax as a way to strip unused modules in
production because tree shaking isn't enabled by default. Tree shaking
isn't enabled by default because `experimentalImportSupport` (ESM
handling) has a number of bugs such as inability to support require
cycles. We need `experimentalImportSupport` because Hermes + RN don't
have built-in support for ESM. Point here is that it'll take a while
before this import/require resolution is a sensible default, so we need
a workaround in the meantime.

# Test Plan

- `npx expo` in an SDK 53 project works as expected.
- Using `EXPO_ATLAS=1 npx expo` shows that only one copy of React
Navigation is pulled in.