[docs] Update EAS Hosting "Worker Runtime" page with Node.js built-in module table (#35899)
# Why

Currently, this page feels a little hard to follow and needed some
updates for several Node.js built-in modules. However, reading and
updating bullet point lists for all Node.js modules is tougher than it
needs to be, and a table would give us more space to elaborate on
implementation notes.

Node.js built-in modules in `workerd` and EAS Hosting are not on a clear
"yes/no" (supported/unsupported) spectrum. Instead, we need some space
to elaborate how a module is supported, if it's missing some
functionality, or if it's not fully implemented. We also need a way to
express that some modules are neither clearly supported or unsupported
i.e. show them as an "inbetween" state that warrants extra attention
from users.

<details>
<summary>Preview Screenshot (does not reflect two latest
commits)</summary>


![image](https://github.com/user-attachments/assets/042a1f3c-ff38-4143-96c3-f7eb98499fd5)

</details>

# How

- Add a new circle alert icon to be used alongside `YesIcon` and
`NoIcon`
  - Note: It's not a perfect match in size, but close enough imho
- Replace lists of supported modules with a table
- Make sure that this table contains all Node.js built-in modules we
account for
- Add implementation notes where appropriate and mark modules that
aren't clearly supported or unsupported with the `AlertIcon` rather than
a `YesIcon` or `NoIcon`
- Replace globals list with a table as well
- Rewrite notes after the updated content

The globals table is not comprehensive (as per WinterTC) but it contains
globals that we're modifying ourselves, i.e. that don't function as
users may expect them to.

# Test Plan

- Ran the docs locally and checked the layout

---------

Co-authored-by: Kadi Kraman <kadi@expo.io>
Co-authored-by: Aman Mittal <amandeepmittal@live.com>