[ios][expo-go] removed processing from UIBackgroundModes
When building expo we run an expo-tools script to remove UIBackgroundModes that shouldn't be included in Expo-Go.

We missed the new `expo-background-task` value `processing`, which caused an exception when validating with the AppStore:

```
The Info.plist key 'BGTaskSchedulerPermittedIdentifiers' must contain a list of identifiers used to submit and handle tasks when 'UIBackgroundModes' has a value of 'processing'.
```

This commit fixes this by also removing the `processing` value from the UIBackgroundModes in the `et eas remove-background-permissions-from-info-plist` command.