[go][android][1/n] Roll out edge-to-edge -  Enable edge-to-edge for all Expo Go projects (#35718)
# Why

ENG-15058
Starting with SDK 53 we want all Expo Go projects to use egde-to-edge

# How

Enabled the edge-to-edge using the `enableEdgeToEdge` function for
expo-go activities.
This works only for projects, which aren't using the `<StatusBar/>`
component, as including it breaks the edge-to-edge (work in progress)

# Test Plan

Tested in Expo Go on Android. Only on a single screen, as I couldn't get
NCL to work.
| App using edge-to-edge (the status bar tint was added manually, by
default it's transparent) | Error Activity in edge-to-edge |
| :---:   | :---: | 
|
![edge-to-edge-app](https://github.com/user-attachments/assets/929b9b2c-1f3a-4709-b354-27029ab64e90)
|
![edge-to-edge-error](https://github.com/user-attachments/assets/46d1c797-2568-44e7-8196-d7c941ff353b)