[docs] change grouping, and page names, in unversioned reference docs (#35824)
# Why

A concept of new pages arrangement and naming in the reference part of
the docs.

Opening this one as a draft, to signify experimental nature of the
changed, and to get a bit more feedback around.

# How

This changes applies only to the `unversioned` docs, and adds a new
dedicated "Third party libraries" packages group, for lightweight
package for 3rd-party libraries that are included in Expo Go default
libraries bundle.

In addition to new section, those pages will now use the pnm package
name as a title, to make it more obvious, which packages we take about,
in comparison to the custom call-names we used so far.

# Test Plan

The changes have been reviewed by running docs app locally. All lint
checks and tests are passing.

# Preview

![Screenshot 2025-04-07 at 15 48
19](https://github.com/user-attachments/assets/7a2be848-cef9-469a-9f36-a3e91175998f)

![Screenshot 2025-04-07 at 15 48
52](https://github.com/user-attachments/assets/a8968426-812d-47fd-b60b-2803cbdc97af)

---------

Co-authored-by: Aman Mittal <amandeepmittal@live.com>